MAIN METHOD -HOTEL MANAGEMENT SYSTEM
This file includes main method from where the execution get started.This also includes HCONTROLLER class which includes details of the categories of the items involve in the menu. This includes "show menulist" method where the customer selects are order he wants to have .
After that main method is declared which includes  creation of different objects in reference to various classes.
Different methods are called using various created objects in the order of the operations that take place in HOTEL.

This whole code is kept in a package called MAIN.
various other packages are also imported in this program.